# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DriveFile {
  id: string;
  name: string;
  mimeType: string;
  size?: number;
  modifiedTime: string;
  webViewLink?: string;
  iconLink?: string;
}

export interface DriveFileListResponse {
  files: DriveFile[];
  total: number;
  configured: boolean;
  message?: string;
}

export interface SyncResponse {
  success: boolean;
  added: number;
  message: string;
}

export interface MergeRequest {
  /** @minItems 1 */
  fileIds: string[];
  /** Desired filename for the merged PDF (without extension) */
  outputName: string;
}

export interface MergeResponse {
  success: boolean;
  /** Download token valid for 10 minutes */
  token?: string;
  fileName?: string;
  pageCount?: number;
  error?: string;
}