# lib/api-zod/src/generated/types/driveFileListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DriveFile } from './driveFile';

export interface DriveFileListResponse {
  files: DriveFile[];
  total: number;
  configured: boolean;
  message?: string;
}