# lib/api-zod/src/generated/types/mergeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MergeRequest {
  /** @minItems 1 */
  fileIds: string[];
  /** Desired filename for the merged PDF (without extension) */
  outputName: string;
}