# lib/api-zod/src/generated/types/mergeResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MergeResponse {
  success: boolean;
  /** Download token valid for 10 minutes */
  token?: string;
  fileName?: string;
  pageCount?: number;
  error?: string;
}